Add Grape::Middleware::Versioner::Base and use Grape::Util::Registry
